requirements.txt
google-generativeai==0.8.3
streamlit==1.38.0
PyMuPDF==1.24.11
reportlab==4.2.2
python-dotenv==1.0.1
pillow==10.4.0
